chore: project folders reorganized
