+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { ProductsComponent } from './products/products.component';
-import { StatisticsComponent } from './statistics/statistics.component';
-import { CoupensComponent } from './coupens/coupens.component';
-import { PagesComponent } from './pages/pages.component';
-import { WeatherComponent } from './weather/weather.component';
-import { SettingsComponent } from './settings/settings.component';
+import { DashboardComponent } from './components/dashboard/dashboard.component';
+import { ProductsComponent } from './components/products/products.component';
+import { StatisticsComponent } from './components/statistics/statistics.component';
+import { CoupensComponent } from './components/coupens/coupens.component';
+import { PagesComponent } from './components/pages/pages.component';
+import { WeatherComponent } from './components/weather/weather.component';
+import { SettingsComponent } from './components/settings/settings.component';
 
 const routes: Routes = [
   {path: '', redirectTo: 'dash', pathMatch: 'full'},
