+++ src/app/components/pages/pages.component.ts
